worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.51.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
]

[tool.ruff]
extend = "~/.config/ruff/ruff.toml"
src = ["src"]
